Hasura webhook for request authorization
The webhook will authorize a request if the value of the `x-hasura-role`
header is found in the currently logged in user's user info external
permissions. A 200 Ok response will be sent in this case together with
the Hasura session data.

In all other cases the request will not be authorized and a 401
Unauthorized response will be sent.