Retry creating the OIDC wiremock server upon failure
In the CI pipeline, wiremock creation has been seen failing, because
wiremock could not bind to the port specified.

Since the affected test run was not the first in the series, it can be
concluded that an earlier test run's port release had not been finished
when the new wiremock creation was invoked.

Therefore, retry creating the wiremock server several times in case
creation fails. This should prevent similar failures in the future.

Using a dynamic wiremock port would be even more failproof. But if the
port is allocated dynamically, the OIDC discovery mechanism could not be
included in the tests (since it is configured via a property containing
the OIDC server's port).