Data source configuration via azure env secrets
Before this change, the data source was configured via the spring
properties spring.datasource.url, ...username, and ...password . The URL
parameter did not allow for easy configuration via azure env secrets,
because the database url parameters are stored in separate secrets
(hostname, name) and also the schema to be used should be separately
configurable.

Therefore, this commit adds manual data source creation to the
application, allowing all parameters to be passed separately. For more
information on which properties are available, see the README.md .

Note that also after this change, no data source is created if the
spring.session.store-type -property is set to "none". This allows the
integration tests to be run without setting up a database.